polish syntax, refactor into new components
